Build river plots for annotations

Build river plots for annotations

build_river_plot(
  anno,
  grouping,
  show_labels = TRUE,
  label_pos = "center",
  min_link_size = 0,
  pad = 0.1,
  fill_group = NULL
)

Arguments

anno

The sample annotations to use. Must have _id, _label, and _color columns for each grouping.

show_labels

Logical, whether or not to show labels. Default is TRUE.

label_pos

Label position - "left", "center", or "right". Can be specified for each entry in group_by. Default = "center".

min_link_size

Numeric, the minimum fraction of cells in either group that must be included in a link for display. Default is 0 (show all links).

pad

The fraction of vertical space to use as padding between groups. Default = 0.1.

fill_group

One group to use as a source for ribbon colors. Default is NULL.

group_by

The bases to use for the river plot, from left to right.

Value

A ggplot2 plot object.