Dot-plot Heatmap plots of group summary statistics

Dot-plot Heatmap plots of group summary statistics

group_dot_plot(
  data,
  anno,
  genes,
  grouping,
  group_order = NULL,
  fill_stat = "median",
  size_stat = "prop_gt0",
  max_size = 10,
  log_scale = TRUE,
  normalize_rows = FALSE,
  colorset = NULL,
  font_size = 7,
  label_height = 25,
  show_counts = TRUE,
  rotate_counts = FALSE,
  max_width = 10,
  return_type = "plot"
)

Arguments

data

A data frame containing gene expression values. The first column should be sample_name

anno

Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns

genes

A character vector containing gene symbols to be plotted.

grouping

A character vector specifying the desc base that should be used to group cells

group_order

Optional: Explicit specification of group order by supplying a vector of group_ids.

fill_stat

The statistic to apply to each group for use as dot fill color. Default = "median". Options are:

"median"

"mean"

"tmean" (25% trimmed mean)

"nzmean" (mean of non-zero values)

"nzmedian" (median of non-zero values)

"prop_gt0" (proportion of samples > 0)

"prop_gt1" (proportion of samples > 1)

"min"

"max"

size_stat

The statistic to apply to each group for scaling dot size. Same options as fill_stat. Default = "prop_gt0".

max_size

Maximum size of dots, in pts.

log_scale

Logical , determines if data is log scaled before plotting. Default = FALSE.

normalize_rows

Logical, whether or not to rescale data within each row of the plot. Default = FALSE.

font_size

numeric object, the font size (in pts) used to make the plot.

label_height

numeric object, Percent of the plot height that should be used for the labels (0 to 100). Default is 25.

show_counts

Logical, whether or not to display sample counts at the top of labels. Default = TRUE.

rotate_counts

Logical, whether or not to rotate sample counts by 90 degrees. Default = FALSE.

max_width

numeric object, percent of plot width that should be used for maximum expression values (0 to 100). Default is 10.

return_type

What values to return - can be "plot", "data", or "both". Default is "plot".

Value

a ggplot2 plot object